block bike crud if not authenticated

+++ server/controllers/bikeControllers.js
@@ -15,19 +15,25 @@ export async function getBikes(req, res) {
 export async function createBike(req, res) {
     const bike = req.body
 
+    if (!req.userId) return res.json({ message: 'Unauthenticated' })
+
     const newBike = new BikeModel(bike)
     try {
         await newBike.save()
+
         // successful creation
-        res.status(201).json(newBike)
+        return res.status(201).json(newBike)
     } catch (error) {
-        res.status(409).json({ message: error.message })
+        return res.status(409).json({ message: error.message })
     }
 }
 
 export async function updateBike(req, res) {
     const { id: _id } = req.params
     const updatedBike = req.body
+
+    if (!req.userId) return res.json({ message: 'Unauthenticated' })
+
     if (!mongoose.Types.ObjectId.isValid(_id)) return res.status(404).send('No post with that id')
 
     const updateResponse = await BikeModel.findByIdAndUpdate(_id, { ...updatedBike, _id }, { new: true })
@@ -37,6 +43,9 @@ export async function updateBike(req, res) {
 
 export async function deleteBike(req, res) {
     const { id: _id } = req.params
+
+    if (!req.userId) return res.json({ message: 'Unauthenticated' })
+
     if (!mongoose.Types.ObjectId.isValid(_id)) return res.status(404).send('No post with that id')
 
     await BikeModel.findByIdAndRemove(_id)
